move flow fill
